+++ src/Componentes/historico/stylesHistorico.js
@@ -1,29 +1,7 @@
-import React from 'react';
-import { View, Text, StyleSheet, ImageBackground, Image, TouchableOpacity } from 'react-native';
+import {StyleSheet} from 'react-native';
 
-import imagem from '../../assets/Job.png';
+export default styles = StyleSheet.create({
 
-export default function EmpregoList(props) {
-
-    return ( 
-        <View style={styles.containerEmpr}>
-            <ImageBackground style={styles.imagemFundo}>
-                <Image 
-                    source={imagem}
-                />
-            </ImageBackground>
-            <View style={styles.detalhesContainer}>
-                <View style={styles.styleFundo}>
-                    <Text style={styles.Texto}>{props.empresa}</Text>
-                    <Text style={styles.Texto}>{props.servico}</Text>
-                </View>
-            </View>
-        </View>
-    )
-}
-2
-
-const styles = StyleSheet.create({
     containerEmpr: {
         justifyContent: 'flex-start',
         alignItems: 'center',
