+++ src/Componentes/tagInput/tagInputSelecionado.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
 
-export default function EmpregoList(props) {
+export default function tagInputSelecionado(props) {
 
     function removerProfissao(nome) {
         props.removeProfissao(nome);
